Converted the Primary Nav somewhat. Still not done though.
