+++ resources/js/Components/Logo.vue
@@ -1,0 +1,7 @@
+<template>
+
+</template>
+
+<script setup>
+
+</script>
