+++ resources/js/Components/SpecialLogo.vue
@@ -1,0 +1,210 @@
+<template>
+    <div id="logo-container" class="lg:h-[3.96vw] h-[7.92vw] lg:w-[33.85vw] w-[67.70vw] relative items-center overflow-hidden ml-[10vw]  rounded-md">
+        <div id="logo-background" class="absolute z-[98] w-full h-full block p-4 top-0 left-0 bg-white rounded-lg shadow-lg border-2 border-white transition-all duration-250 ease-in-out">
+            <div class="ripple"></div>
+        </div>
+        <div class="overflow-hidden lg:h-[3.96vw] h-[7.92vw]">
+            <div id="logo-background2"
+                 class="absolute z-[98] lg:h-[3.96vw] h-[7.92vw] lg:w-[33.85vw] w-[67.70vw] block top-0 left-0 bg-transparent rounded-lg shadow-lg border-2 border-white transition-all duration-250 ease-in-out">
+            </div>
+            <div class="relative">
+                <svg id="joebeze_logo_lazer_anim" class="absolute lg:h-[3.96vw] h-[7.92vw] lg:w-[33.85vw] w-[67.70vw] z-[99] top-[0.68vw] left-[0.73vw]" data-name="joebeze_logo_lazer_anim"
+                     xmlns="http://www.w3.org/2000/svg" viewBox="0 0 650 76.27">
+                    <defs>
+                        <style>
+                            .cls-1 {
+                                fill: transparent;
+                                stroke-width: 0px;
+                            }
+                        </style>
+                    </defs>
+                    <path class="cls-1"
+                          d="M26.82,4.59v24.75c0,7.56-3.98,13.05-14,13.05H0v-6.61h12.43c5.94,0,7.73-2.86,7.73-6.5V4.59h6.66Z"/>
+                    <path class="cls-1"
+                          d="M52.92,12.82c7.84,0,12.82,4.82,12.82,14.78s-4.98,14.78-12.82,14.78h-6.22c-7.9,0-12.82-4.82-12.82-14.78s4.93-14.78,12.82-14.78h6.22Z"/>
+                    <path class="cls-1"
+                          d="M47.26,18.54c-5.15,0-7.39,2.69-7.39,9.07s2.24,9.02,7.39,9.02h5.1c5.1,0,7.39-2.63,7.39-9.02s-2.3-9.07-7.39-9.07h-5.1Z"/>
+                    <path class="cls-1"
+                          d="M92.12,12.82c7,0,9.97,3.36,9.97,9.07s-2.74,8.51-10.02,8.51h-14.06v2.13c0,2.35,1.29,4.09,4.54,4.09h18.09v5.77h-18.09c-7.39,0-10.3-4.14-10.3-9.74v-10.08c0-5.6,2.91-9.74,10.3-9.74h9.58Z"/>
+                    <path class="cls-1"
+                          d="M78.01,25.76h13.89c3.19,0,4.37-1.34,4.37-3.53s-1.29-3.7-4.2-3.7h-9.52c-3.25,0-4.54,1.79-4.54,4.09v3.14Z"/>
+                    <path class="cls-1"
+                          d="M115.3,0v12.82h13.1c7.84,0,12.77,4.82,12.77,14.78s-4.93,14.78-12.77,14.78h-8.18c-7.11,0-10.86-4.14-10.86-12.66V0h5.94Z"/>
+                    <path class="cls-1"
+                          d="M115.3,18.54v10.92c0,5.04,1.85,7.17,6.66,7.17h5.82c5.15,0,7.39-2.63,7.39-9.02s-2.24-9.07-7.39-9.07h-12.49Z"/>
+                    <path class="cls-1"
+                          d="M167.55,12.82c7,0,9.97,3.36,9.97,9.07s-2.74,8.51-10.02,8.51h-14.06v2.13c0,2.35,1.29,4.09,4.54,4.09h18.09v5.77h-18.09c-7.39,0-10.3-4.14-10.3-9.74v-10.08c0-5.6,2.91-9.74,10.3-9.74h9.58Z"/>
+                    <path class="cls-1"
+                          d="M153.44,25.76h13.89c3.19,0,4.37-1.34,4.37-3.53s-1.29-3.7-4.2-3.7h-9.52c-3.25,0-4.54,1.79-4.54,4.09v3.14Z"/>
+                    <path class="cls-1"
+                          d="M204.01,12.82c5.15,0,7.62,2.69,7.62,6.78,0,3.64-1.96,5.99-5.04,7.11l-16.97,6.16c-1.01.39-1.57,1.06-1.57,2.02,0,1.12.73,1.79,2.24,1.79h20.72v5.71h-21.06c-5.1,0-7.62-2.69-7.62-6.78,0-3.3,1.62-5.94,5.04-7.17l17.02-6.1c1.23-.45,1.57-1.29,1.57-2.07,0-1.12-.73-1.74-2.3-1.74h-20.27v-5.71h20.61Z"/>
+                    <path class="cls-1"
+                          d="M237.27,12.82c7,0,9.97,3.36,9.97,9.07s-2.74,8.51-10.02,8.51h-14.06v2.13c0,2.35,1.29,4.09,4.54,4.09h18.09v5.77h-18.09c-7.39,0-10.3-4.14-10.3-9.74v-10.08c0-5.6,2.91-9.74,10.3-9.74h9.58Z"/>
+                    <path class="cls-1"
+                          d="M223.16,25.76h13.89c3.19,0,4.37-1.34,4.37-3.53s-1.29-3.7-4.2-3.7h-9.52c-3.25,0-4.54,1.79-4.54,4.09v3.14Z"/>
+                    <path class="cls-1"
+                          d="M299.93,0v29.74c0,8.51-3.81,12.66-10.86,12.66h-8.18c-7.9,0-12.82-4.82-12.82-14.78s4.93-14.78,12.82-14.78h13.05V0h5.99Z"/>
+                    <path class="cls-1"
+                          d="M293.94,29.46v-10.92h-12.49c-5.1,0-7.39,2.69-7.39,9.07s2.3,9.02,7.39,9.02h5.88c4.76,0,6.61-2.13,6.61-7.17Z"/>
+                    <path class="cls-1"
+                          d="M327.65,12.82c7,0,9.97,3.36,9.97,9.07s-2.74,8.51-10.02,8.51h-14.06v2.13c0,2.35,1.29,4.09,4.54,4.09h18.09v5.77h-18.09c-7.39,0-10.3-4.14-10.3-9.74v-10.08c0-5.6,2.91-9.74,10.3-9.74h9.58Z"/>
+                    <path class="cls-1"
+                          d="M313.54,25.76h13.89c3.19,0,4.37-1.34,4.37-3.53s-1.29-3.7-4.2-3.7h-9.52c-3.25,0-4.54,1.79-4.54,4.09v3.14Z"/>
+                    <path class="cls-1"
+                          d="M346.97,12.82l9.86,23.74c.22.62.73.9,1.46.9.78,0,1.29-.28,1.57-.9l10.25-23.74h6.38l-11.42,26.38c-1.23,2.8-3.42,3.86-6.78,3.86s-5.6-1.12-6.72-3.86l-10.86-26.38h6.27Z"/>
+                    <path class="cls-1"
+                          d="M402.69,0v12.82h13.1c7.84,0,12.77,4.82,12.77,14.78s-4.93,14.78-12.77,14.78h-8.18c-7.11,0-10.86-4.14-10.86-12.66V0h5.94Z"/>
+                    <path class="cls-1"
+                          d="M402.69,18.54v10.92c0,5.04,1.85,7.17,6.66,7.17h5.82c5.15,0,7.39-2.63,7.39-9.02s-2.24-9.07-7.39-9.07h-12.49Z"/>
+                    <path class="cls-1" d="M441.61,0v42.39h-5.94V0h5.94Z"/>
+                    <path class="cls-1"
+                          d="M467.76,12.82c7.84,0,12.82,4.82,12.82,14.78s-4.98,14.78-12.82,14.78h-6.22c-7.9,0-12.82-4.82-12.82-14.78s4.93-14.78,12.82-14.78h6.22Z"/>
+                    <path class="cls-1"
+                          d="M462.11,18.54c-5.15,0-7.39,2.69-7.39,9.07s2.24,9.02,7.39,9.02h5.1c5.1,0,7.39-2.63,7.39-9.02s-2.3-9.07-7.39-9.07h-5.1Z"/>
+                    <path class="cls-1"
+                          d="M507.3,12.82c7.06,0,10.86,4.14,10.86,12.66v17.02c0,8.4-2.91,11.7-11.2,11.7h-17.58v-5.71h16.91c4.42,0,5.88-1.74,5.88-5.99v-1.34h-13.05c-7.9,0-12.82-4.65-12.82-14.22s4.93-14.11,12.82-14.11h8.18Z"/>
+                    <path class="cls-1"
+                          d="M499.68,35.78h12.49v-10.36c0-4.82-1.85-6.89-6.61-6.89h-5.88c-5.1,0-7.39,2.58-7.39,8.62s2.3,8.62,7.39,8.62Z"/>
+                </svg>
+            </div>
+            <div class="relative">
+                <svg id="joebeze_logo_filled" class="absolute lg:h-[3.96vw] h-[7.92vw] lg:w-[33.85vw] w-[67.70vw] z-[99] top-[0.47vw] left-[0.78vw]" data-name="joebeze_logo_filled" xmlns="http://www.w3.org/2000/svg"
+                     viewBox="0 0 650 76.27">
+                    <defs>
+                        <style> <!--The 2nd svg classes are all cls-2-x ---->
+                        .cls-2-1 {
+                            fill: url('#tessPattern');
+                            font-family: NasalizationRg-Regular, Nasalization;
+                            font-size: 56px;
+                            transition: fill 2s ease;
+                        }
+
+                        .cls-2-2 {
+                            letter-spacing: -.01em;
+                        }
+
+                        .cls-2-3 {
+                            letter-spacing: -.02em;
+                        }
+
+                        .cls-2-4 {
+                            letter-spacing: -.02em;
+                        }
+                        </style>
+                        <pattern id="tessPattern" patternUnits="userSpaceOnUse" width="256" height="256">
+                            <image href="{{ asset('images/bytebeze-banner-tess.png') }}" width="256"
+                                   height="256" preserveAspectRatio="xMidYMid slice"/>
+                        </pattern>
+                        <pattern id="fillPattern" patternUnits="userSpaceOnUse" width="100%" height="100%">
+                            <image href="{{ asset('images/Screenshot 2024-03-23 100050.jpg') }}" width="100%"
+                                   height="100%" preserveAspectRatio="xMidYMid slice"/>
+                        </pattern>
+                    </defs>
+                    <text class="cls-2-1" transform="translate(0 46.9)">
+                        <tspan x="0" y="0">Joeb</tspan>
+                        <tspan class="cls-2-3" x="145.65" y="0">e</tspan>
+                        <tspan class="cls-2-2" x="181.1" y="0">z</tspan>
+                        <tspan x="215.37" y="0">e d</tspan>
+                        <tspan class="cls-2-4" x="305.76" y="0">e</tspan>
+                        <tspan x="341.26" y="0">v blog</tspan>
+                    </text>
+                </svg>
+            </div>
+            <div class="relative">
+                <svg id="ripple-effect-bg" class="z-20 absolute lg:top-[-18.33vw] top-[-36.66vw]  lg:left-[-4.17vw] left-[-8.34vw] w-[98.96vw] lg:w-[49.48vw] h-[83.34vw] lg:h-[41.67vw] rounded-md shadow-lg border-2 border-white transition-all duration-1000 ease-in-out"
+                     data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" viewBox="0 0 1070.78 901.04">
+                    <defs>
+                        <style>
+                            .cls-3-1 {
+                                clip-path: url(#clippath-2);
+                            }
+
+                            .cls-3-2 {
+                                fill: none;
+                                stroke-width: 0px;
+                            }
+
+                            .cls-3-3 {
+                                clip-path: url(#clippath-1);
+                            }
+
+                            .cls-3-4 {
+                                clip-path: url(#clippath-4);
+                            }
+
+                            .cls-3-5 {
+                                clip-path: url(#clippath);
+                            }
+
+                            .cls-3-6 {
+                                clip-path: url(#clippath-3);
+                            }
+
+                            .cls-3-7 {
+                                clip-path: url(#clippath-5);
+                            }
+                        </style>
+                        <clipPath id="clippath">
+                            <circle id="_EllipseWhite10_" data-name="&lt;EllipseWhite10&gt;" class="cls-3-2" cx="-100" cy="-100" r="1"/>
+                        </clipPath>
+                        <clipPath id="clippath-1">
+                            <circle class="cls-3-2" cx="-100" cy="-100" r="1"/>
+                        </clipPath>
+                        <!--            <clipPath id="clippath-2">-->
+                        <!--                <circle class="cls-3-2" cx="-100" cy="-100" r="1"/>-->
+                        <!--            </clipPath>-->
+                        <clipPath id="clippath-3">
+                            <circle class="cls-3-2" cx="-100" cy="-100" r="1"/>
+                        </clipPath>
+                        <clipPath id="clippath-4">
+                            <circle class="cls-3-2" cx="-100" cy="-100" r="1"/>
+                        </clipPath>
+                        <clipPath id="clippath-5">
+                            <circle class="cls-3-2" cx="-100" cy="-100" r="1"/>
+                        </clipPath>
+                    </defs>
+                    <g id="_Clip_Group_Redhead_" data-name="&lt;Clip Group Redhead&gt;">
+                        <g class="cls-3-5">
+                            <image id="redhead" width="1017" height="256" transform="translate(80.85 340) scale(.76)"/>
+                        </g>
+                    </g>
+                    <g id="_Clip_Group_Blueninja_" data-name="&lt;Clip Group Blueninja&gt;">
+                        <g class="cls-3-3">
+                            <image id="blueninja" width="1017" height="256" transform="translate(80.85 340) scale(.76)"/>
+                        </g>
+                    </g>
+                    <!--        <g id="_Clip_Group_Darkgame_" data-name="&lt;Clip Group Darkgame&gt;">-->
+                    <!--            <g class="cls-3-1">-->
+                        <!--                <image width="640" height="360" transform="translate(150.31 277.3) scale(1.05)" xlink:href="{{ asset('images/dark-game.gif') }}"/>-->
+                        <!--            </g>-->
+                    <!--        </g>-->
+                    <g id="_Clip_Group_Blackcat_" data-name="&lt;Clip Group Blackcat&gt;">
+                        <g class="cls-3-6">
+                            <image id="blackcat" width="1017" height="256" transform="translate(80.85 331.18) scale(.76)"/>
+                        </g>
+                    </g>
+                    <g id="_Clip_Group_Tanwoman_" data-name="&lt;Clip Group Tanwoman&gt;">
+                        <g class="cls-3-4">
+                            <image id="tanwoman" width="1017" height="256" transform="translate(80.85 331.18) scale(.76)"/>
+                        </g>
+                    </g>
+                    <g id="_Clip_Group_Orangewoman_" data-name="&lt;Clip Group Orangewoman&gt;">
+                        <g class="cls-3-7">
+                            <image id="orangewoman" width="1017" height="256" transform="translate(80.85 331.18) scale(.76)"/>
+                        </g>
+                    </g>
+                </svg>
+            </div>
+
+
+            <canvas id="canvas" class="absolute top-0 left-0 lg:h-[3.96vw] h-[7.92vw] lg:w-[33.85vw] w-[67.70vw]"></canvas>
+            <canvas id="canvas2" class="z-[99] absolute top-0 left-0 lg:h-[3.96vw] h-[7.92vw] lg:w-[33.85vw] w-[67.70vw]"></canvas>
+        </div>
+    </div>
+</template>
+
+<script setup>
+
+
+
+</script>
